+++ src/pages/InputEntryPage.jsx
@@ -5,18 +5,17 @@ import 'react-calendar/dist/Calendar.css';
 import '../style/InputEntryPage.scss';
 import moment from 'moment';
 import backbutton from '../IMG/backbutton.png';
+import InputEntryModal from '../pages/InputEntryModal'
 
 const InputEntryPage = () => {
   const { date } = useParams();
   const navigate = useNavigate();
   const selectedDate = new Date(date);
 
-  const [category, setCategory] = useState('');
-  const [customCategory, setCustomCategory] = useState('');
-  const [amount, setAmount] = useState('');
-  const [storeName, setStoreName] = useState(''); // 명세서 대로 추가
-  const [description, setDescription] = useState(''); // 명세서 대로 추가 22
+  const [isAddModalOpen, setIsAddModalOpen] = useState(false);
 
+  const openAddModal = () => setIsAddModalOpen(true);
+  const closeAddModal = () => setIsAddModalOpen(false);
   // 날짜 클릭 시 해당 날짜 페이지로 이동
   const handleDateClick = (newDate) => {
     const year = newDate.getFullYear();
@@ -35,39 +34,6 @@ const InputEntryPage = () => {
   const handleBackClick = () => {
     navigate('/main');
   };
-
-  const handleSave = async () => {
-    const selectedCategory = category === '기타' ? customCategory : category;
-    if (!selectedCategory || !amount) {
-      alert('카테고리와 금액을 모두 입력해주세요!');
-      return;
-    }
-  
-    const expenseData = {
-      expenseCategory: selectedCategory,
-      amount: parseInt(amount, 10),
-      storeName: storeName,  // 가게 상호명 
-      description: description  // 메모 
-    };
-  
-    try {
-      const response = await axios.post(`http://백엔드서버주소/expense/${date}`, expenseData, {
-        headers: {
-          'Content-Type': 'application/json',
-        },
-      });
-  
-      if (response.data.success) {
-        alert('지출이 성공적으로 저장되었습니다!');
-        navigate('/main');  // 저장 후 메인 페이지로 이동
-      } else {
-        alert('지출 저장에 실패했습니다.');
-      }
-    } catch (error) {
-      console.error('지출 저장 오류:', error);
-      alert('서버와 통신 중 문제가 발생했습니다.');
-    }
-  }; // 백엔드로 보내는 수동 지출 입력 로직
 
   return (
     <div className="input-entry-page">
@@ -99,66 +65,18 @@ const InputEntryPage = () => {
             prev2Label={null}
             minDetail="year"
           />
-        </div>
 
-        <div className="form-button-wrapper">
-          <div className="entry-form">
-            <select
-              value={category}
-              onChange={(e) => setCategory(e.target.value)}
-              className="category-select"
-            >
-              <option value="">카테고리를 선택하세요</option>
-              <option value="식비">🍽️ 식비</option>
-              <option value="교통비">🚗 교통비</option>
-              <option value="쇼핑">🛍️ 쇼핑</option>
-              <option value="주거비">🏠 주거비</option>
-              <option value="취미/여가">🎨 취미/여가</option>
-              <option value="기타">✏️ 기타</option>
-            </select>
-
-            {category === '기타' && (
-              <input
-                type="text"
-                placeholder="직접 입력하세요"
-                value={customCategory}
-                onChange={(e) => setCustomCategory(e.target.value)}
-                className="custom-category-input"
-              />
-            )}
-
-            <input
-              type="number"
-              placeholder="금액을 입력하세요 (₩)"
-              value={amount}
-              onChange={(e) => setAmount(e.target.value)}
-              className="amount-input"
-            />
-
-            <input
-              type="text"
-              placeholder="상호명을 입력하세요 "
-              value={storeName}
-              onChange={(e) => setStoreName(e.target.value)}
-              className="storename-input"
-            />
-
-            <input
-              type="text"
-              placeholder=" 메모를 추가하세요 "
-              value={description}
-              onChange={(e) => setDescription(e.target.value)}
-              className="description-input"
-            />
-
-            <button className="save-button" onClick={handleSave}>
-              저장하기
+          <div className="form-button-wrapper">
+            <button className="add-expense-button" onClick={openAddModal}>
+              👉 직접 추가하기
+            </button>
+            <button className="receipt-button" onClick={handleReceiptPage}>
+              🧾 영수증으로 추가하기
             </button>
           </div>
-
-          <button className="receipt-button" onClick={handleReceiptPage}>
-            🧾
-          </button>
+        </div>
+        <div>
+          <InputEntryModal isOpen={isAddModalOpen} onClose={closeAddModal} />
         </div>
       </div>
     </div>
